+++ .storybook/main.js
@@ -1,4 +1,4 @@
 module.exports = {
-  stories: ['../packages/**/*.stories.{ts,tsx,js,jsx}'],
+  stories: ['../packages/**/*.stories.{ts,tsx,js,jsx,mdx}'],
   addons: ['@storybook/addon-links', '@storybook/addon-actions', '@storybook/addon-essentials'],
 };