+++ .storybook/preview.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { ThemeProvider } from 'styled-components';
 
-import '@reach/dialog/styles.css';
-
 import { theme } from '../packages/zephyr/src/theme';
 
 export const parameters = {