+++ packages/zephyr/src/theme/variants/index.ts
@@ -1,0 +1,10 @@
+import { button } from './button';
+import { text } from './text';
+
+export const variants = {
+  ...button,
+  ...text,
+};
+
+export * from './button';
+export * from './text';